+++ Kosmobook/users/urls.py
@@ -1,4 +1,6 @@
 from django.urls import path
+from .views import profile_view
 
 urlpatterns = [
+    path('<str:username>/', profile_view, name='profile'),
 ]
