--- users/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import profile_view
-
-urlpatterns = [
-    path('<str:username>/', profile_view, name='profile'),
-]
